Refactor RateLimiter rules into separate files
Refactored the `IRateLimitRule`, `FixedWindowRule`, `SlidingWindowRule`, and `RegionalRule` classes into their own files within the `RateLimiter.Rules` namespace to improve code organization and modularity. Updated `RateLimiterTest.cs` and `RateLimiter.cs` to reflect these changes. Added additional region checks for "ASIA" and a default "GLOBAL" region in the `RegionalRule` class.
