+++ RateLimiter/Rules/IRateLimitRule.cs
@@ -1,0 +1,8 @@
+﻿namespace RateLimiter.Rules
+{
+    public interface IRateLimitRule
+    {
+        bool IsAllowed(string clientId, string resource);
+        void Cleanup();
+    }
+}
